fix(generate): remove unnecessary log statements to reduce clutter refactor(domain): move Verify method to validators.go for better organization style(generate): clean up extra newlines and whitespace
