+++ tools/seltabls/domain/validators.go
@@ -3,6 +3,9 @@ package domain
 import (
 	"fmt"
 	"net/url"
+	"strings"
+
+	"github.com/PuerkitoBio/goquery"
 )
 
 // IsValidGoType checks if the given type is a valid go type
@@ -85,6 +88,52 @@ func ValidateConfig(cfg *ConfigFile) error {
 	}
 	if cfg.FastModel == "" {
 		return fmt.Errorf("fast-model is required")
+	}
+	return nil
+}
+
+// Verify checks if the selectors are in the html
+func (f *Field) Verify(htmlBody string) error {
+	doc, err := goquery.NewDocumentFromReader(
+		strings.NewReader(htmlBody),
+	)
+	if err != nil {
+		return fmt.Errorf("failed to create document: %w", err)
+	}
+	if f.DataSelector != "" {
+		sel := doc.Find(f.DataSelector)
+		if sel.Length() == 0 {
+			return fmt.Errorf("failed to find selector: %s", f.DataSelector)
+		}
+	} else {
+		return fmt.Errorf("no data found for selector %s with type %s in field %s with type %s", f.DataSelector, f.Type, f.Name, f.Type)
+	}
+	if f.ControlSelector != "" {
+		sel := doc.Find(f.ControlSelector)
+		if sel.Length() == 0 {
+			return fmt.Errorf("failed to find selector: %s", f.ControlSelector)
+		}
+	} else {
+		return fmt.Errorf("no control found for selector %s with type %s in field %s with type %s", f.ControlSelector, f.Type, f.Name, f.Type)
+	}
+	if f.QuerySelector != "" {
+		sel := doc.Find(f.QuerySelector)
+		if sel.Length() == 0 {
+			return fmt.Errorf("failed to find selector: %s", f.QuerySelector)
+		}
+	} else {
+		return fmt.Errorf("no query found for selector %s with type %s in field %s with type %s", f.QuerySelector, f.Type, f.Name, f.Type)
+	}
+	if f.HeaderSelector != "" {
+		sel := doc.Find(f.HeaderSelector)
+		if sel.Length() == 0 {
+			return fmt.Errorf("failed to find selector: %s", f.HeaderSelector)
+		}
+	}
+	mbp := f.MustBePresent
+	docTxt := doc.Text()
+	if !strings.Contains(docTxt, mbp) {
+		return fmt.Errorf("must be present (%s) not found for field %s with type %s", mbp, f.Name, f.Type)
 	}
 	return nil
 }
